move artists subquery to user resolver
